refactor(core): remove getRecoveryFeePerBytes
Replace with MempoolApi in `utxo/recovery`

Issue: BG-38773
